Update init.ijs
This is debugged code outside of whisk using a docker image and curl into a JHS platform. Right now you must use jijx to set up the init and run apps. The Dockerfile will need to be changed to add a configuration file or a J profile script so everything is 
pre-setup in JHS at start up.