Update run.ijs
Added a catch block around the call to the main script. This is for start up purposes in case run gets called before a main script has been loaded. It also catches any errors that might occur in the script so an error JSON message can be sent back. Also call the main script from the jhs namespace where init loads it.